requirements.txt
openai-whisper
deep-translator